Crate stake

Stake

Minimal staking pallet that implements ordered validator selection by total amount at stake

Rules

There is a new round every BlocksPerRound blocks.

At the start of every round,

IssuancePerRound is distributed to validators for BondDuration rounds ago in proportion to the points they received in that round (for authoring blocks)

queued validator exits are executed

a new set of validators is chosen from the candidates

To join the set of candidates, an account must call join_candidates with stake >= MinValidatorStk and fee <= MaxFee. The fee is taken off the top of any rewards for the validator before the remaining rewards are distributed in proportion to stake to all nominators (including the validator, who always self-nominates).

To leave the set of candidates, the validator calls leave_candidates. If the call succeeds, the validator is removed from the pool of candidates so they cannot be selected for future validator sets, but they are not unstaked until BondDuration rounds later. The exit request is stored in the ExitQueue and processed BondDuration rounds later to unstake the validator and all of its nominators.

To join the set of nominators, an account must call join_nominators with stake >= MinNominatorStk. There are also runtime methods for nominating additional validators and revoking nominations.

Structs

Bond

GenesisConfig
Genesis config for the module, allow to build genesis storage.

InflationInfo

Module

Nominator

Range

Validator
Global validator state with commission fee, bonded stake, and nominations

ValidatorSnapshot
Snapshot of validator state at the start of the round for which they are selected

Enums

Call
Dispatchable calls.

Error

RawEvent
Events for this module.

ValidatorStatus
The activity status of the validator

Traits

Config

_
Provides version information about a pallet.

Type Definitions

Event
RawEvent specialized for the configuration Config